refactor: restructure app-context location
Moved `app-context` from `contexts` to `components/layout/app-context`. Updated all relevant imports for consistency and improved folder structure. This enhances the modularity and organization of shared components.
