+++ apps/nextjs/src/components/layout/app-context/global-query.tsx
